+++ packages/web/config/analytics-events.ts
