+++ packages/web/pages/stake.tsx
@@ -14,8 +14,7 @@ import { Spinner } from "~/components/loaders/spinner";
 import { UnbondingInProgress } from "~/components/stake/unbonding-in-progress";
 import { StakeOrUnstake } from "~/components/types";
 import { StakeOrEdit } from "~/components/types";
-import { EventName } from "~/config";
-import { AmountDefault } from "~/config/user-analytics-v2";
+import { AmountDefault, EventName } from "~/config";
 import { useAmountConfig, useFakeFeeConfig } from "~/hooks";
 import { useAmplitudeAnalytics, useGetApr, useTranslation } from "~/hooks";
 import { useStakedAmountConfig } from "~/hooks/ui-config/use-staked-amount-config";
